Creat rs-fMRI folder and T1 folder
Why:

Put naren's project here

This change addresses the need by:

All
