+++ rs-fMRI/v2.0/models/ABCD_rr_models_base.sav
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0728bb258dd7b746f1443f2b0e64ed7734842e5a2d365fec5547054b611e01f3
+size 25375970
