+++ rs-fMRI/v2.0/models/HBN_rr_models_base.sav
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:0019ed774a1e25e99a99faa6ffdc6da5824acaee8074e0ddc093641e58469b5b
+size 29603561
