--- v2.0/models/ABCD_rr_models_multilayer.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:8b89b974c0f8070851345b53ddb08260192097d9643510c8edb0c963cb2df960
-size 232850373
