--- v2.0/models/GSP_rr_models_base.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:e1ad976a2ad37a77ea407169413c5e4bba4d6969b22f3c0633e36c3159d1ceb6
-size 16211605
