--- v2.0/models/GSP_rr_models_multilayer.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:56c7242470e844a57c3f38d428a8ad8ddaf1cbcd0119ad649cbde3020c9cef7a
-size 27219234
