--- v2.0/models/HBN_rr_models_multilayer.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:a21208561e8cd32a775aec65c09b127f368e95d8c0a507c9f84ef526a88a4ec5
-size 49726364
