--- v2.0/models/UKBB_rr_models.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:9d8101a5fc75a02bb0a6a0c66ceaddaa864d809e0e8134fd9002f1e8681da298
-size 47223340
