--- v2.0/models/eNKI_rr_models_base.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:6e80554f04377863972b31185ad43d43d6f3a37a20af1b0a556908516d1e80a7
-size 42994552
