--- v2.0/models/eNKI_rr_models_multilayer.sav
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:d3d76257109c2e96a0ee8414af08a8e3ef53c6465097cf3464d9712711b8c1c8
-size 72949382
